Split parsing and processing
This includes a lru_cache for the result of the parsing, which should
help with any performance concerns with parsing the statements on
every usage.

Note the mypy tests fail as there is much work to do to improve and
finalise this code. Note also the parser currently only accepts max
two word SQL commands.