Better support expression groups
This will allow for queries with (x = 1 AND y = {y}) where y is Absent
to be correctly rewritten as (x = 1) as this is now recognised as a
group of expressions rather than a singular expression as before.

Also note separators are better represented now as a specific (suffix)
aspect of an expression.